Index + About section overhaul
Started to change docs overall to reflect what MS does more accurately, as opposed to the more abstract explanation of metadata vs data generation. Everything will eventually follow the three step pipeline: estimation, (de)serialization and generation.

Images have been added to illustrate this pipeline.

The index has been updated in line with this, in addition, it also now has an overview of key features.

The about section is in progress of being overhauled, with the intention of housing all the technical detail there. A lot of text here is sourced from the WIP JOSS paper.

I have also made some slight changes to the extensions page, and metadata generation page (changed it to metaframe generation) but will continue to make changes in the following commits.